Apply fixes from StyleCI

+++ src/Valkyrja/ORM/Entities/Entity.php
@@ -261,11 +261,11 @@ abstract class Entity extends Model implements Contract
      * @param string $property The property name
      * @param mixed  $value    The value
      *
-     * @return double
+     * @return float
      */
     protected function __getDoubleValueForDataStore(string $property, mixed $value): float
     {
-        return (double) $value;
+        return (float) $value;
     }
 
     /**
